refactor notification tests
